fix(vmmdgrpc): bound-check Seccomp_filters before int32 conversion
Closes the CodeQL 'Incorrect conversion between integer types' alert
on the new ParseSeccompLines function (alert #N, fail row in the
CodeQL rollup on PR #237 — surfaced post-rebase when the action's
analysis re-ran on the merge ref).

filterLen was previously returned as 'int' (architecture-dependent
bit width) and then narrowed at the protobuf assignment site
(FilterLen: int32(filterLen)) without bounds checking. CodeQL
flags the unchecked narrowing as a potential truncation bug.

Fix: ParseSeccompLines and readSeccompStatus now return int32
directly. The Seccomp_filters source is bounded against
math.MaxInt32 before the narrowing — the kernel reports the count
of attached BPF programs (in practice 1, max observed in this repo
is 1 for the jailer's policy). A value that doesn't fit int32 is
either a kernel bug or a forged /proc, so we fail loud rather
than silently truncate to a wrong value.

The e2e test wrapper (parseSeccomp in cmd/e2e) was updated to the
matching int32 return.

Why this slipped past the original PR #237 CI: the rebased branch
+ new main both run CodeQL on the merge ref, and the go-language
analysis re-evaluates against the post-rebase state. The
actions/unpinned-tag fix is in place (commit a383dde); this fixes
the follow-on go/incorrect-integer-conversion alert that surfaced
in the same rollup.

+++ pkg/vmmdgrpc/seccomp_test.go
@@ -90,7 +90,7 @@ func TestParseSeccompLines_AllKnownModes(t *testing.T) {
 			if mode != tc.wantMode {
 				t.Errorf("mode = %q, want %q", mode, tc.wantMode)
 			}
-			if filter != tc.wantFilter {
+			if filter != int32(tc.wantFilter) {
 				t.Errorf("filter = %d, want %d", filter, tc.wantFilter)
 			}
 		})
